+++ packages/brick-container/index.d.ts
@@ -56,6 +56,24 @@ export interface DevConfig {
   userConfigByApps?: UserConfigByApps;
 
   /**
+   * 启用 https
+   *
+   * @example
+   * ```js
+   * export default {
+   *   https: {
+   *     key: readFileSync("./localhost.key.pem", "utf-8"),
+   *     cert: readFileSync("./localhost.cert.pem", "utf-8"),
+   *   },
+   * }
+   * ```
+   */
+  https?: {
+    key: string;
+    cert: string;
+  };
+
+  /**
    * API mocks
    *
    * @example
@@ -78,6 +96,11 @@ export interface DevConfig {
    * ```
    */
   mocks?: RequestHandler[];
+
+  /**
+   * 过滤需要进行 size-check 的构件。
+   */
+  sizeCheckFilter?: (brick: string, pkgId?: string) => boolean;
 }
 
 interface Settings {