+++ packages/brick-container/serve/env.js
@@ -106,6 +106,8 @@ export async function getEnv(rootDir, runtimeFlags) {
   const devConfigMjs = path.join(rootDir, "dev.config.mjs");
   let configuredBrickFolders = false;
   let userConfigByApps;
+  let https;
+  let sizeCheckFilter;
   if (existsSync(devConfigMjs)) {
     const devConfig = (await import(devConfigMjs)).default;
     if (devConfig) {
@@ -116,6 +118,8 @@ export async function getEnv(rootDir, runtimeFlags) {
       localSettings = devConfig.settings;
       localMocks = devConfig.mocks;
       userConfigByApps = devConfig.userConfigByApps;
+      https = devConfig.https;
+      sizeCheckFilter = devConfig.sizeCheckFilter;
     }
   }
 
@@ -147,11 +151,13 @@ export async function getEnv(rootDir, runtimeFlags) {
     liveReload: flags.liveReload,
     localSettings,
     userConfigByApps,
+    https,
     host: flags.host,
     port: Number(flags.port),
     wsPort: Number(flags.wsPort),
     server: getServerPath(flags.server),
     sizeCheck: flags.sizeCheck,
+    sizeCheckFilter,
     verbose: flags.verbose,
   };