+++ packages/brick-container/serve/utils/sizeCheck.js
@@ -8,7 +8,7 @@ export function getSizeCheckApp() {
   };
 }
 
-export function getSizeCheckStoryboards(brickPackages) {
+export function getSizeCheckStoryboards(brickPackages, sizeCheckFilter) {
   return [
     {
       app: getSizeCheckApp(),
@@ -21,12 +21,18 @@ export function getSizeCheckStoryboards(brickPackages) {
             {
               brick: "ul",
               children: brickPackages.flatMap((pkg) =>
-                pkg.bricks.concat(pkg.elements ?? []).map((brick) => ({
-                  brick: "li",
-                  properties: {
-                    textContent: `${pkg.id.split("/")[1]}:${brick}`,
-                  },
-                }))
+                pkg.bricks
+                  .concat(pkg.elements ?? [])
+                  .filter(
+                    (brick) =>
+                      !sizeCheckFilter || sizeCheckFilter(brick, pkg.id)
+                  )
+                  .map((brick) => ({
+                    brick: "li",
+                    properties: {
+                      textContent: `${pkg.id.split("/")[1]}:${brick}`,
+                    },
+                  }))
               ),
             },
             {
@@ -38,17 +44,22 @@ export function getSizeCheckStoryboards(brickPackages) {
           ],
         },
         ...brickPackages.flatMap((pkg) =>
-          pkg.bricks.concat(pkg.elements ?? []).map((brick) => ({
-            path: `\${APP.homepage}/${brick}`,
-            exact: true,
-            preLoadBricks: [brick],
-            bricks: [
-              {
-                brick: "p",
-                properties: { textContent: "This is size-check!" },
-              },
-            ],
-          }))
+          pkg.bricks
+            .concat(pkg.elements ?? [])
+            .filter(
+              (brick) => !sizeCheckFilter || sizeCheckFilter(brick, pkg.id)
+            )
+            .map((brick) => ({
+              path: `\${APP.homepage}/${brick}`,
+              exact: true,
+              preLoadBricks: [brick],
+              bricks: [
+                {
+                  brick: "p",
+                  properties: { textContent: "This is size-check!" },
+                },
+              ],
+            }))
         ),
 
         // By each package
@@ -76,12 +87,11 @@ export function getSizeCheckStoryboards(brickPackages) {
         ...brickPackages.map((pkg) => ({
           path: `\${APP.homepage}/packages/${pkg.id.split("/")[1]}`,
           exact: true,
-          preLoadBricks: pkg.bricks.concat(pkg.elements ?? []).filter(
-            (brick) =>
-              // This brick is an alias of a deprecated brick.
-              // Ignore it otherwise it will cause custom element conflict.
-              pkg.id !== "bricks/basic" || brick !== "basic.app-bar-wrapper"
-          ),
+          preLoadBricks: pkg.bricks
+            .concat(pkg.elements ?? [])
+            .filter(
+              (brick) => !sizeCheckFilter || sizeCheckFilter(brick, pkg.id)
+            ),
           bricks: [
             {
               brick: "p",
@@ -96,12 +106,7 @@ export function getSizeCheckStoryboards(brickPackages) {
           exact: true,
           preLoadBricks: brickPackages
             .flatMap((pkg) => pkg.bricks.concat(pkg.elements ?? []))
-            .filter(
-              (brick) =>
-                // This brick is an alias of a deprecated brick.
-                // Ignore it otherwise it will cause custom element conflict.
-                brick !== "basic.app-bar-wrapper"
-            ),
+            .filter((brick) => !sizeCheckFilter || sizeCheckFilter(brick)),
           bricks: [
             {
               brick: "p",