+++ packages/easyops-runtime/src/flowApi/CollectContracts.ts
@@ -2,6 +2,7 @@ import { Contract } from "@next-core/types";
 
 const contractsMap: Map<string, Contract> = new Map();
 const widgetContractMap: Map<string, Contract> = new Map();
+const debugContractMap: Map<string, Contract> = new Map();
 
 const addContract = (
   contracts: Contract[] | undefined,
@@ -24,6 +25,18 @@ export function clearCollectWidgetContract(): void {
   widgetContractMap.clear();
 }
 
+export function collectDebugContract(contracts: Contract[] | undefined): void {
+  addContract(contracts, debugContractMap);
+}
+
+export function clearDebugContract() {
+  debugContractMap.clear();
+}
+
 export function getContract(name: string): Contract | undefined {
-  return contractsMap.get(name) || widgetContractMap.get(name);
+  return (
+    contractsMap.get(name) ||
+    widgetContractMap.get(name) ||
+    debugContractMap.get(name)
+  );
 }