+++ packages/utils/src/general/createSSEStream.spec.ts
@@ -156,12 +156,10 @@ describe("createSSEStream", () => {
       for await (const _value of iterator) {
         // Do nothing
       }
-    }).rejects.toMatchInlineSnapshot(
-      `[SyntaxError: Unexpected token 'E', "[ERROR]: Oops" is not valid JSON]`
-    );
+    }).rejects.toMatchInlineSnapshot(`[Error: [ERROR]: Oops]`);
     expect(consoleError).toHaveBeenCalledTimes(1);
     expect(consoleError).toHaveBeenCalledWith(
-      "Failed to parse message data:",
+      "Received error message:",
       "[ERROR]: Oops"
     );
   });
@@ -174,7 +172,7 @@ describe("createSSEStream", () => {
       await Promise.resolve();
       onmessage?.({ id: "", event: "", data: '"1"' });
       await Promise.resolve();
-      onmessage?.({ id: "", event: "", data: "[ERROR]: Oops" });
+      onmessage?.({ id: "", event: "", data: "Oops" });
       await Promise.resolve();
       onmessage?.({ id: "", event: "", data: "[DONE]" });
     });
@@ -186,12 +184,55 @@ describe("createSSEStream", () => {
         // Do nothing
       }
     }).rejects.toMatchInlineSnapshot(
-      `[SyntaxError: Unexpected token 'E', "[ERROR]: Oops" is not valid JSON]`
+      `[SyntaxError: Unexpected token 'O', "Oops" is not valid JSON]`
     );
     expect(consoleError).toHaveBeenCalledTimes(1);
     expect(consoleError).toHaveBeenCalledWith(
-      "Failed to parse message data:",
-      "[ERROR]: Oops"
+      "Failed to parse message:",
+      "Oops"
     );
+  });
+
+  test("error event", async () => {
+    mockFetchEventSource.mockImplementation(async (_url, options) => {
+      const { onopen, onmessage } = options;
+      await Promise.resolve();
+      await onopen?.({ ok: true } as Response);
+      await Promise.resolve();
+      onmessage?.({ id: "", event: "", data: '"1"' });
+      await Promise.resolve();
+      onmessage?.({ id: "", event: "error", data: '{"error":"Oops"}' });
+    });
+    const iterator = await createSSEStream<string>(url);
+    await expect(async () => {
+      for await (const _value of iterator) {
+        // Do nothing
+      }
+    }).rejects.toMatchInlineSnapshot(`[Error: Oops]`);
+    expect(consoleError).toHaveBeenCalledTimes(1);
+    expect(consoleError).toHaveBeenCalledWith(
+      "Received error event:",
+      '{"error":"Oops"}'
+    );
+  });
+
+  test("error event that can not be parsed as JSON", async () => {
+    mockFetchEventSource.mockImplementation(async (_url, options) => {
+      const { onopen, onmessage } = options;
+      await Promise.resolve();
+      await onopen?.({ ok: true } as Response);
+      await Promise.resolve();
+      onmessage?.({ id: "", event: "", data: '"1"' });
+      await Promise.resolve();
+      onmessage?.({ id: "", event: "error", data: "Yaks" });
+    });
+    const iterator = await createSSEStream<string>(url);
+    await expect(async () => {
+      for await (const _value of iterator) {
+        // Do nothing
+      }
+    }).rejects.toMatchInlineSnapshot(`[Error: Yaks]`);
+    expect(consoleError).toHaveBeenCalledTimes(1);
+    expect(consoleError).toHaveBeenCalledWith("Received error event:", "Yaks");
   });
 });