+++ packages/build-next-libs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.15](https://github.com/easyops-cn/next-core/compare/@next-core/build-next-libs@1.0.14...@next-core/build-next-libs@1.0.15) (2024-05-10)
+
+**Note:** Version bump only for package @next-core/build-next-libs
+
+
+
+
+
 ## [1.0.14](https://github.com/easyops-cn/next-core/compare/@next-core/build-next-libs@1.0.13...@next-core/build-next-libs@1.0.14) (2024-04-16)
 
 **Note:** Version bump only for package @next-core/build-next-libs