+++ packages/test-next/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/easyops-cn/next-core/compare/@next-core/test-next@1.1.0...@next-core/test-next@1.1.1) (2024-05-10)
+
+**Note:** Version bump only for package @next-core/test-next
+
+
+
+
+
 # [1.1.0](https://github.com/easyops-cn/next-core/compare/@next-core/test-next@1.0.17...@next-core/test-next@1.1.0) (2024-05-09)